# backend/Swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel & Table Management API",
        "version": "1.0.0",
        "description": "API for managing hotels and tables"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        }
    ],
    "paths": {
...
        "/owner/create": {
            "post": {
                "summary": "Create a new sport and associated slots",
                "tags": [
                    "Sport Owner"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSportRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sport and slots created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateSportResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/get_user_details": {
            "post": {
                "summary": "Get user details by email",
                "tags": [
                    "Sport Owner"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "required": [
                                    "email"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/booking-details/{userId}": {
            "get": {
                "summary": "Get all booking details for a user",
                "tags": [
                    "Sport Owner"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "User ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns user details, bookings, and wallet"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/get_user_bookings": {
            "post": {
                "summary": "Get sport bookings for a user by email",
                "tags": [
                    "Sport Owner"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "required": [
                                    "email"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns list of booked sports"
                    },
                    "404": {
                        "description": "No bookings found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateSportRequest": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "cost": {
                        "type": "number"
                    },
                    "address": {
                        "type": "string"
                    },
                    "contact_mail": {
                        "type": "string"
                    },
                    "slot_timings": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "required": [
                    "label",
                    "body",
                    "cost",
                    "address",
                    "contact_mail",
                    "slot_timings",
                    "date"
                ]
            },
            "CreateSportResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "sport": {
                        "$ref": "#/components/schemas/Sport"
                    },
                    "bookings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Booking"
                        }
                    }
                }
            },
...
            "Sport": {
                "type": "object",
                "required": [
                    "label",
                    "body",
                    "cost",
                    "address",
                    "contact_mail"
                ],
                "properties": {
                    "label": {
                        "type": "string",
                        "maxLength": 30,
                        "pattern": "^[A-Za-z]+$"
                    },
                    "body": {
                        "type": "string",
                        "description": "Must contain at least 5 letters and only valid characters"
                    },
                    "cost": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 10000
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 200,
                        "description": "Must contain at least 4 letters and only valid characters"
                    },
                    "contact_mail": {
                        "type": "string",
                        "format": "email"
                    }
                }
            },
...
            "Booking": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "theatrename": {
                        "type": "string"
                    },
                    "screenNo": {
                        "type": "string"
                    },
                    "movie": {
                        "type": "string"
                    },
                    "theatre": {
                        "type": "string"
                    },
                    "screen": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "timeSlot": {
                        "type": "string"
                    },
                    "selectedSeatCodeMap": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    },
                    "totalPrice": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        },
...
    }
}